refactor: first shot at using `RouterService`
