+++ addon/instance-initializers/ember-router-scroll.js
@@ -1,0 +1,8 @@
+export function initialize(appInstance) {
+  // Eagerly initialize service
+  appInstance.lookup('service:router-scroll');
+}
+
+export default {
+  initialize
+};
