+++ app/instance-initializers/ember-router-scroll.js
@@ -1,0 +1,1 @@
+export { default, initialize } from 'ember-router-scroll/instance-initializers/ember-router-scroll';
