+++ tests/dummy/app/router.js
@@ -1,7 +1,7 @@
-import RouterScroll from 'ember-router-scroll';
+import EmberRouter from '@ember/routing/router';
 import config from './config/environment';
 
-class Router extends RouterScroll {
+class Router extends EmberRouter {
   location = config.locationType;
   rootURL = config.rootURL;
 }
